+++ docs/changelog.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## Version 0.5 [2024-04-21]
+
+- [#119](https://github.com/guibranco/pancake/issues/119) - Add OneSignal class by [@guibranco](https://github.com/guibranco).
+
 ## Version 0.4 [2024-04-21]
 
-- [#115](https://github.com/guibranco/pancake/issues/1510) - Add GUID (v4) class by [@guibranco](https://github.com/guibranco).
+- [#115](https://github.com/guibranco/pancake/issues/115) - Add GUID (v4) class by [@guibranco](https://github.com/guibranco).
 
 ## Version 0.3 [2024-04-08]